refactor(api): move constants for challenge's status and type to shared and move BaseChallenge to evaluation
