+++ api/tests/tooling/domain-builder/factory/evaluation/index.js
@@ -1,7 +1,9 @@
+import { buildBaseChallenge } from './build-base-challenge.js';
 import { buildChallengeToPlay } from './build-challenge-to-play.js';
 import { buildSmartRandomChallenge } from './build-smart-random-challenge.js';
 
 export const builders = {
+  buildBaseChallenge,
   buildSmartRandomChallenge,
   buildChallengeToPlay,
 };
