+++ api/tests/tooling/domain-builder/factory/learning-content/build-challenge.js
@@ -1,4 +1,5 @@
-import { Challenge, STATUSES, TYPES } from '../../../../../src/learning-content/domain/models/Challenge.js';
+import { Challenge } from '../../../../../src/learning-content/domain/models/Challenge.js';
+import { CHALLENGE_STATUSES as STATUSES, CHALLENGE_TYPES as TYPES } from '../../../../../src/shared/constants.js';
 
 export const buildChallenge = function buildChallenge({
   id = 'foo id',
