--- api/tests/tooling/domain-builder/factory/shared/index.js
@@ -1,5 +1,0 @@
-import { buildBaseChallenge } from './build-base-challenge.js';
-
-export const builders = {
-  buildBaseChallenge,
-};
